Work with the most recent versions of Chrome for testing.
Update how we reference local tiles to work with the latest versions of
Chrome.  Specifically, using a proxy and http (rather than https)
doesn't work, even when http goes to localhost or 127.0.0.1.  This seems
to be a defect in Chrome since version 68.  This changes the test url
for tiles to the local server, which avoids the problem.